Add towncrier configuration
towncrier is a utility to produce useful, summarized news files (also
known as changelogs) for your project [1]. It provides us with the
ability for contributor to write their own meaningful additions to
the changelog without risking continuous merge conflicts.

This should be used for v2. v1 can continue to use the legacy method.

[1] https://pypi.org/project/towncrier/

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>