CI: Switch to Makefile for running acceptance tests
We can remove the 'acceptancetest' script in the process. Note that this
isn't a 1:1 match: the fwaasv2 test job now runs the entire networking
test suite and not just the fwaasv2-specific tests but seeing as this is
using OVS rather than OVN that doesn't seem like such a bad thing.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>